Attempt 1 at cleanup
