+++ src/qlearning/smallExampleQAgent.py
